requirements.txt
absl-py==0.9.0
alembic==1.4.2
altair==4.1.0
appdirs==1.4.4
asn1crypto==0.24.0
astor==0.8.1
astroid==2.4.2
astunparse==1.6.3
atpublic==1.0
attrs==19.3.0
azure-core==1.7.0
azure-storage-blob==12.3.2
backcall==0.2.0
base58==2.0.1
bleach==3.1.5
blinker==1.4
boto3==1.14.7
botocore==1.17.7
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cloudpickle==1.5.0
colorama==0.4.3
configobj==5.0.6
configparser==5.0.0
cryptography==2.1.4
databricks-cli==0.11.0
decorator==4.4.2
defusedxml==0.6.0
distro==1.5.0
docker==4.2.2
docker-pycreds==0.4.0
docutils==0.15.2
dpath==2.0.1
dvc==0.94.1
entrypoints==0.3
enum-compat==0.0.3
Flask==1.1.2
flatten-json==0.1.7
flufl.lock==3.2
funcy==1.14
future==0.18.2
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.3
google-auth==1.14.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
gorilla==0.3.0
gql==0.2.0
grandalf==0.6
graphql-core==1.1
grpcio==1.28.1
gunicorn==20.0.4
h5py==2.10.0
humanize==2.4.0
idna==2.6
importlib-metadata==1.6.1
inflect==3.0.2
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.11.2
jmespath==0.10.0
jsonpath-ng==1.5.1
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
lazy-object-proxy==1.4.3
Mako==1.1.3
Markdown==3.2.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
mlflow==1.9.1
msrest==0.6.17
nanotime==0.5.2
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.3
notebook==6.0.3
numpy==1.18.4
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opt-einsum==3.2.1
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.0
pathspec==0.8.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
ply==3.11
prometheus-client==0.8.0
prometheus-flask-exporter==0.15.0
promise==2.3
prompt-toolkit==3.0.5
protobuf==3.11.3
psutil==5.7.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycrypto==2.6.1
pydeck==0.4.0b2
pydot==1.4.1
Pygments==2.6.1
pygobject==3.26.1
pygtrie==2.3.2
pylint==2.5.3
pyparsing==2.4.7
pyrsistent==0.16.0
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
pyxdg==0.25
PyYAML==5.3.1
pyzmq==19.0.1
querystring-parser==1.2.4
requests==2.23.0
requests-oauthlib==1.3.0
rope==0.17.0
rsa==4.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
scipy==1.4.1
SecretStorage==2.3.1
Send2Trash==1.5.0
sentry-sdk==0.15.1
shortuuid==1.0.1
six==1.14.0
smmap==3.0.4
SQLAlchemy==1.3.13
sqlparse==0.3.1
streamlit==0.61.0
subprocess32==3.5.4
tabulate==0.8.7
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
texttable==1.6.2
toml==0.10.1
toolz==0.10.0
tornado==5.1.1
tqdm==4.46.1
traitlets==4.3.3
treelib==1.6.1
typed-ast==1.4.1
tzlocal==2.1
urllib3==1.25.9
validators==0.15.0
voluptuous==0.11.7
wandb==0.9.1
watchdog==0.10.2
wcwidth==0.2.4
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zc.lockfile==2.0
zipp==3.1.0
